panoptic/epsnet2/requirements.txt
pillow<7.0
tensorboardX
numpy
umsgpack
torchvision==0.3
tqdm
scikit-image
pycocotools
git+git://github.com/waspinator/pycococreator.git@0.2.0
git+https://github.com/mapillary/inplace_abn.git
